% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{check_list}
\alias{check_list}
\title{Check list of data scanning}
\format{
A nested list with fishery name (the top level) and each column names
which this pkg assumes as input.
}
\usage{
check_list
}
\description{
An item list about purse-seine fishery logbook.
}
\keyword{datasets}
