% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_ideal_colnames}
\alias{fetch_ideal_colnames}
\title{Genarater of ideal column names}
\usage{
fetch_ideal_colnames(fishery)
}
\arguments{
\item{fishery}{fishery type of logbook}
}
\value{
a sequence of name strings having to be equipped
}
\description{
Genarater of ideal column names
}
\examples{
fetch_ideal_colnames(fishery = "purse seine")

}
