% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%all_in\%}
\alias{\%all_in\%}
\title{All value matching}
\usage{
x \%all_in\% X
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{X}{vector or NULL: the values to be matched against.}
}
\value{
TRUE or FALSE
}
\description{
All value matching
}
\examples{
letters[1:3] \%all_in\% letters[1:5]
letters[1:3] \%all_in\% letters[2:5]
}
