% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-is.R
\name{is_all_uniq_vals_in}
\alias{is_all_uniq_vals_in}
\title{Test numbers to be correctly digitalized}
\usage{
is_all_uniq_vals_in(dat, column, u_vals)
}
\arguments{
\item{dat}{data frame including column to be inspected}

\item{column}{a column name ("string") to be scaned}

\item{u_vals}{unique values being expected to include all target values}
}
\value{
an warning message
}
\description{
Test numbers to be correctly digitalized
}
\examples{
\dontrun{
is_all_uniq_vals_in(dat, column = "foo", u_vals = 1:2)
}
}
