% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-is.R
\name{is_sum}
\alias{is_sum}
\title{Test total value corresponded with the sum of partial values}
\usage{
is_sum(dat, column, ...)
}
\arguments{
\item{dat}{data frame including column to be inspected}

\item{column}{a column name ("string") of total value digitized}

\item{...}{a sequence of column names to be sumed up}
}
\value{
an warning message
}
\description{
Test total value corresponded with the sum of partial values
}
\examples{
data <- data.frame(a = 1:3, b = 4:6, c = 7:9, sum = 1:3 + 4:6 + 7:9)
is_sum(data, column = "sum", "a", "b", "c")
is_sum(data, column = "sum", 1:3)
is_sum(data, column = "sum", c("a", "b", "c"))
is_sum(data, column = "sum", `a`:`c`)
}
