% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inner.R
\name{is_sum_equal}
\alias{is_sum_equal}
\title{Test total value corresponded with the sum of partial values}
\usage{
is_sum_equal(dat, total_col, ...)
}
\arguments{
\item{dat}{data frame including column to be inspected}

\item{total_col}{a column name ("string") of total value digitized}

\item{...}{a sequence of column names to be sumed up}
}
\value{
an warning message
}
\description{
Test total value corresponded with the sum of partial values
}
\examples{
data <- data.frame(a = 1:3, b = 4:6, c = 7:9, sum = 1:3 + 4:6 + 7:9)
is_sum_equal(data, total_col = "sum", "a", "b", "c")
is_sum_equal(data, total_col = "sum", 1:3)
is_sum_equal(data, total_col = "sum", c("a", "b", "c"))
is_sum_equal(data, total_col = "sum", `a`:`c`)

}
