% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inner.R
\name{is_unique_values}
\alias{is_unique_values}
\title{Test a number to be correctly digitalized}
\usage{
is_unique_values(dat, column, u_vals)
}
\arguments{
\item{dat}{data frame including column to be scaned}

\item{column}{a column name ("string") to be scaned}

\item{u_vals}{unique values to be matched against}
}
\value{
any warning messages
}
\description{
Test a number to be correctly digitalized
}
\examples{
\dontrun{
is_unique_values(dat, column = "foo", u_vals = 2020)
}
}
