% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{post_isFunc}
\alias{post_isFunc}
\title{Classify is_functions}
\usage{
post_isFunc(type = 1, ...)
}
\arguments{
\item{type}{a string or a number. Must be one of 1 to 4, or
\code{c("uniq_len", "uniq_val", "all_in", "sum")}; then, it calls
\code{is_unique_length}, \code{is_unique_values},
\code{is_all_uniq_vals_in}, and \code{is_sum}, respectively.}

\item{...}{arguments of each called function.}
}
\value{
TRUE of FALSE
}
\description{
Classify is_functions
}
\examples{
\dontrun{
post_isFunc(type = "uniq_len")
post_isFunc(type = 1)
}
}
