% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-by-row.R
\name{scan_by_row}
\alias{scan_by_row}
\title{Detect row(s) which is/are suspected}
\usage{
scan_by_row(.d)
}
\arguments{
\item{.d}{a data.frame}
}
\value{
a list of
}
\description{
Detect row(s) which is/are suspected
}
\examples{
\dontrun{
scan_by_row(df)
}
}
