% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-contents.R
\name{scan_contents}
\alias{scan_contents}
\title{Inspect and suggest changes contents of input data}
\usage{
scan_contents(.dat, species, dictionaries = NULL, .file, .sheet, .fishery)
}
\arguments{
\item{.dat}{a data.frame or tibble standardized by \code{uniform_df()}.}

\item{species}{species such as "not_tuna"}

\item{dictionaries}{lookup tables referred to.}

\item{.file}{file name}

\item{.sheet}{sheet name when excel file read}

\item{.fishery}{fishery such as "purse_seine"}
}
\value{
a data.frame
}
\description{
Inspect and suggest changes contents of input data
}
\examples{
\dontrun{
scan_contents()
}
}
