% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_rm_newline_code}
\alias{str_rm_newline_code}
\title{Remove all newline codes in any strings}
\usage{
str_rm_newline_code(vector)
}
\arguments{
\item{vector}{vector to be processed}
}
\value{
Cleaned up string(s)
}
\description{
Remove all newline codes in any strings
}
\examples{
x <- c("Main:\rSub", "Main2:\nSub", "Main3:\r\nSub")
str_rm_newline_code(x)
}
