% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-functions.R
\name{test_excel_logbook}
\alias{test_excel_logbook}
\title{Read an input excel file (logbook) and test it.}
\usage{
test_excel_logbook(
  file,
  sheet,
  fishery = "purse seine",
  species = "not-tunas",
  dictionary = "vessel-license-2020.csv"
)
}
\arguments{
\item{file}{an input excel file}

\item{sheet}{an input sheet in the input file}

\item{fishery}{fishery that the logbook treats}

\item{species}{species that the logbook includes}

\item{dictionary}{a look-up table of vessel name and licence number in
a specific year}
}
\value{
data.frame describing error or warning records
}
\description{
Read an input excel file (logbook) and test it.
}
\examples{
\dontrun{
test_excel_logbooks(file = "your/file.xlsx",
                    sheet = "Sheet 1",
                    fishery = "purse seine",
                    species = "not-tunas",
                    dictionary = "vessel-license-2020.csv")
}
              
}
