% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-functions.R
\name{test_excel_logbooks}
\alias{test_excel_logbooks}
\title{Read input files in a directory (folder) and test them all.}
\usage{
test_excel_logbooks(
  dir,
  sheet,
  fishery = "purse seine",
  species = "not-tunas",
  dictionary = "vessel-license-2020.csv"
)
}
\arguments{
\item{dir}{an input directory which has one or more input files}

\item{sheet}{an input sheet name in common with all input excel files}

\item{fishery}{fishery that the logbook treats}

\item{species}{species that the logbook includes}

\item{dictionary}{a look-up table of vessel name and licence number in
a specific year}
}
\value{
data.frame describing error or warning records
}
\description{
Read input files in a directory (folder) and test them all.
}
\examples{
\dontrun{
test_excel_logbooks(dir = "your/directory",
                    sheet = "Sheet 1",
                    fishery = "purse seine",
                    species = "not-tunas",
                    dictionary = "vessel-license-2020.csv")
}

#' @export
}
