% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inner.R
\name{test_length_identical}
\alias{test_length_identical}
\title{Test length to be correctyly digitalized}
\usage{
test_length_identical(dat, column, u_len)
}
\arguments{
\item{dat}{data frame including column to be scaned}

\item{column}{a column name ("string") to be scaned}

\item{u_len}{an unique value (ideal length) to be matched against}
}
\value{
any warning messages
}
\description{
Test length to be correctyly digitalized
}
\examples{
\dontrun{
test_length_identical(dat, column = "foo", u_len = 1L)
}
}
