% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-functions.R
\name{test_logbook}
\alias{test_logbook}
\title{Read an input excel file (logbook) and test it.}
\usage{
test_logbook(file, fishery, species, dictionary)
}
\arguments{
\item{file}{input file}

\item{fishery}{fishery that the logbook treats}

\item{species}{species that the logbook includes}

\item{dictionary}{a look-up table of vessel name and licence number in a specific year}
}
\value{

}
\description{
Read an input excel file (logbook) and test it.
}
\examples{
\dontrun {
lgbktestr::test_logbooks(file = "your/file.xlsx",
                         fishery = "purse seine",
                         species = "not-tunas",
                         dictionary = "vessel-license-2020.csv")
}
}
