% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-functions.R
\name{test_logbooks}
\alias{test_logbooks}
\title{Read input files in a directory (folder) and test them all.}
\usage{
test_logbooks(dir, fishery, species, dictionary)
}
\arguments{
\item{dir}{an input directory which has one or more input files}

\item{fishery}{fishery that the logbook treats}

\item{species}{species that the logbook includes}

\item{dictionary}{a look-up table of vessel name and licence number in a specific year}
}
\value{

}
\description{
Read input files in a directory (folder) and test them all.
}
\examples{
\dontrun {
test_logbooks(dir = "your/directory",
              fishery = "purse seine",
              species = "not-tunas",
              dictionary = "vessel-license-2020.csv")
}
}
