% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inner.R
\name{test_region_coexistence}
\alias{test_region_coexistence}
\title{Check regions not to be intermingled}
\usage{
test_region_coexistence(dat, column)
}
\arguments{
\item{dat}{data frame including column to be scaned}

\item{column}{a column name to be scaned}
}
\value{
any warning messages
}
\description{
Check regions not to be intermingled
}
\examples{
\dontrun{
test_region_coexistence(dat, column = "操業海域")
}
}
