% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inner.R
\name{test_sum}
\alias{test_sum}
\title{Test total value corresponded with the sum of partial values}
\usage{
test_sum(dat, total_col, partial_cols = list())
}
\arguments{
\item{dat}{data frame including column to be inspected}

\item{total_col}{a column name of total value digitized}

\item{partial_cols}{a list of column names to sum up}
}
\value{
an warning message
}
\description{
Test total value corresponded with the sum of partial values
}
\examples{
\dontrun{
test_sum(dat, total_col = "foobar", partial_cols = list("foo", "bar"))
}
}
