% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inner.R
\name{test_value_equals}
\alias{test_value_equals}
\title{Test numbers to be correctly digitalized}
\usage{
test_value_equals(dat, column, u_values)
}
\arguments{
\item{dat}{data frame including column to be inspected}

\item{column}{a column name to be scaned}

\item{u_values}{unique values (ideal numbers) to be matched against}
}
\value{
an warning message
}
\description{
Test numbers to be correctly digitalized
}
\examples{
\dontrun{
test_region_numcode(dat, column = "foo", ideal_code_numbers = 1:2)
}
}
