% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inner.R
\name{test_value_identical}
\alias{test_value_identical}
\title{Test a number to be correctly digitalized}
\usage{
test_value_identical(dat, column, u_value)
}
\arguments{
\item{dat}{data frame including column to be scaned}

\item{column}{a column name ("string") to be scaned}

\item{u_value}{an unique value (ideal number) to be matched against}
}
\value{
any warning messages
}
\description{
Test a number to be correctly digitalized
}
\examples{
\dontrun{test_year(dat, column = "foo", u_value = 2020)}
}
