% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-inner.R
\name{test_year}
\alias{test_year}
\title{Check digitalized year}
\usage{
test_year(dat, column, year)
}
\arguments{
\item{dat}{data frame including column to be scaned}

\item{column}{a column name to be scaned}

\item{year}{an year to be matched against}
}
\value{
any warning messages
}
\description{
Check digitalized year
}
\examples{
\dontrun{
test_year(dat, column = "foo")
}
}
