% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat-csv.R, R/treat-excel.R
\name{treat_excel}
\alias{treat_excel}
\title{Check up logbook csv file.}
\usage{
treat_excel(
  file,
  sheet,
  fishery = "purse seine",
  species = "not-tunas",
  dictionaries
)

treat_excel(
  file,
  sheet,
  fishery = "purse seine",
  species = "not-tunas",
  dictionaries
)
}
\arguments{
\item{file}{an excel file}

\item{sheet}{an input sheet in the input file}

\item{fishery}{fishery which the logbook treats}

\item{species}{species which the logbook treats}

\item{dictionaries}{a look-up table of vessel name and licence number in
a specific year}
}
\value{
data.frame describing error or warning records

data.frame describing error or warning records
}
\description{
Check up logbook csv file.

Check up logbook excel file.
}
\examples{
\dontrun{
treat_excel(file = "your/file.csv",
            fishery = "purse seine",
            species = "not-tunas",
            dictionaries = list("files/at/", "your/local"))
}
\dontrun{
treat_excel(file = "your/file.xlsx",
            sheet = "Sheet 1",
            fishery = "purse seine",
            species = "not-tunas",
            dictionaries = list("files/at/", "your/local"))
}
}
