% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniform_df}
\alias{uniform_df}
\title{Uniform (Standardize) data frame.}
\usage{
uniform_df(df)
}
\arguments{
\item{df}{input data frame to be processed}
}
\value{
A data.frame ncol reduced if excess columns exists.
}
\description{
Uniform (Standardize) data frame.
}
\examples{
df <- data.frame(整理番号 = 1:3,
                 操業次 = rep(2, 3),
                 報告年 = rep(2019, 3))
uniform_df(df)
 Warnings:
  Removed column(s): '報告年'
}
