% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-functions.R
\name{update_check_list}
\alias{update_check_list}
\title{Update check list}
\usage{
update_check_list(fishery = c(), column = c(), values = c(), isType = c())
}
\arguments{
\item{fishery}{fishery name strings such as "purse_seine", "long_line"
(snakle_case).}

\item{column}{column name strings contained in a target data.frame.}

\item{values}{values to be matched against.}

\item{isType}{Choose from "uniq_len", "uniq_val", "all_in", "sum" use for
\code{is_unique_values}, \code{is_unique_values}, \code{is_all_uniq_vals_in},
\code{is_sum}.}
}
\description{
Update check list
}
