% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-by-row.R
\name{which_false}
\alias{which_false}
\title{Give the \code{FALSE} indices of a logical object, allowing for array indices}
\usage{
which_false(x)
}
\arguments{
\item{x}{a logical vector or array. NAs}
}
\value{
indices (usually number)
}
\description{
Give the \code{FALSE} indices of a logical object, allowing for array indices
}
\examples{
which_false(c(T, T, F, F)) 
# [1] 3 4
}
