% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-contents.R
\name{scan_contents}
\alias{scan_contents}
\title{Inspect data contents}
\usage{
scan_contents(.dat, species, dictionaries = NULL, .file, .sheet, .fishery)
}
\arguments{
\item{.dat}{A data.frame or tibble standardized by \code{uniform_df()}.}

\item{species}{Species such as "not_tuna".}

\item{dictionaries}{Lookup tables referred to.}

\item{.file}{File name.}

\item{.sheet}{Sheet name when excel file read.}

\item{.fishery}{Fishery such as "purse_seine".}
}
\value{
A data.frame.
}
\description{
Scan input data (\code{data.frame}-like object) and return suggestions that
records might be problematic.
}
\examples{
\dontrun{
scan_contents()
}
}
