% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat-excel.R
\name{treat_excel}
\alias{treat_excel}
\title{Check up logbook excel file.}
\usage{
treat_excel(file, sheet, fishery, ...)
}
\arguments{
\item{file}{an excel file}

\item{sheet}{an input sheet in the input file}

\item{fishery}{fishery which the logbook treats}

\item{...}{arguments succeeded to each method function}
}
\value{
data.frame describing error or warning records
}
\description{
Check up logbook excel file.
}
\examples{
\dontrun{
treat_excel(file = "your/file.xlsx",
            sheet = "Sheet 1",
            fishery = "purse_seine"))
}
}
