% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniform_df}
\alias{uniform_df}
\title{Uniform (Standardize) data frame.}
\usage{
uniform_df(df, fishery = "purse seine")
}
\arguments{
\item{df}{data frame to be processed}

\item{fishery}{a fishery type string}
}
\value{
A data.frame ncol reduced if excess columns exists.
}
\description{
Uniform (Standardize) data frame.
}
\examples{
df <- data.frame(A = 1:3,
                 B = rep(2, 3),
                 C = rep(3, 3))
uniform_df(df)

}
